lolcompiler

Function main

pub(crate) fn main()

lolcode compiler

This is a compiler for this grammer

==============================================================================================================

Terminals

HAI ::= “#HAI”

KBYE ::= “#KBYE”

OBTW ::= “#OBTW”

TLDR ::= “#TLDR”

MAEK ::= “#MAEK”

OIC ::= “#OIC”

GIMMEH ::= “#GIMMEH”

MKAY ::= “#MKAY”

HEAD ::= “HEAD”

TITLE ::= “TITLE”

PARAGRAF ::= “PARAGRAF”

BOLD ::= “BOLD”

ITALICS ::= “ITALICS”

LIST ::= “LIST”

ITEM ::= “ITEM”

NEWLINE ::= “NEWLINE”

LINX ::= “LINX”

IHAZ ::= “#IHAZ”

ITIZ ::= “#ITIZ”

LEMMESEE ::= “#LEMMESEE”

VARNAME ::= any single word (A–Z, a–z, no spaces)

VARVALUE ::= allowed text characters

TEXT ::= plain text (letters, digits, punctuation, spaces)

ADDRESS ::= text without spaces

Note that these terminals are not case sensitive.

The only allowed plain text in our language is: A-Z, a-z, 0-9, commas, period, quotes, colons, question marks, exclamation points, percent signs, forward slashes, equals sign, and hidden special characters for tabs and newlines. Syntax/Production Rules (i.e., non-terminals)

::= HAI < head> KBYE

::=

| ε

::= OBTW TEXT TLDR